Backend/current/requirements.txt
gunicorn
psycopg2-binary
Django
django-admin-interface
django-colorfield
django-environ
django-jalali
django-cors-headers
django-rest-knox
djangorestframework
requests